Added SDL2 and new component system

+++ Pacman/Vector2f.h
@@ -81,7 +81,7 @@ public:
 	//FUNCTIONS
 	float Vector2f::Length() const
 	{
-		return sqrt(X * X + Y * Y);
+		return sqrtf(X * X + Y * Y);
 	}
 
 	void Vector2f::Normalize()
